topics/mcp/requirements.txt
fastmcp
openai-agents
python-dotenv
